Refactor Login and LogoutLogin listener, settingsactivity and settingsactivitystrategies.
