+++ app/src/cambodia/java/org/eyeseetea/malariacare/strategies/SettingsActivityStrategy.java
@@ -1,10 +1,8 @@
 package org.eyeseetea.malariacare.strategies;
 
-import android.content.Intent;
 import android.preference.Preference;
 import android.preference.PreferenceScreen;
 
-import org.eyeseetea.malariacare.LoginActivity;
 import org.eyeseetea.malariacare.SettingsActivity;
 import org.eyeseetea.malariacare.layout.listeners.LogoutAndLoginRequiredOnPreferenceClickListener;
 import org.eyeseetea.malariacare.layout.listeners.PullRequiredOnPreferenceChangeListener;
@@ -47,23 +45,7 @@ public class SettingsActivityStrategy extends ASettingsActivityStrategy {
 
     @Override
     public Preference.OnPreferenceChangeListener getOnPreferenceChangeListener() {
-
         return pullRequiredOnPreferenceChangeListener;
     }
 
-
-    @Override
-    public boolean onPreferenceClick(final Preference preference) {
-        if (!settingsActivity.getIntent().getBooleanExtra(SettingsActivity.IS_LOGIN_DONE, false)) {
-
-            String orgUnitValue = settingsActivity.autoCompleteEditTextPreference.getText();
-
-            Intent loginIntent = new Intent(settingsActivity, LoginActivity.class);
-            loginIntent.putExtra(LoginActivity.PULL_REQUIRED, orgUnitValue.isEmpty());
-
-            settingsActivity.startActivity(loginIntent);
-        }
-        return true;
-    }
 }
-
