+++ app/src/laos/java/org/eyeseetea/malariacare/strategies/SettingsActivityStrategy.java
@@ -1,12 +1,10 @@
 package org.eyeseetea.malariacare.strategies;
 
-import android.content.Intent;
 import android.preference.Preference;
 import android.preference.PreferenceScreen;
 
-import org.eyeseetea.malariacare.LoginActivity;
 import org.eyeseetea.malariacare.SettingsActivity;
-import org.eyeseetea.malariacare.layout.listeners.LogoutAndLoginRequiredOnPreferenceClickListener;
+import org.eyeseetea.malariacare.layout.listeners.LoginRequiredOnPreferenceClickListener;
 import org.eyeseetea.malariacare.layout.listeners.PullRequiredOnPreferenceChangeListener;
 
 public class SettingsActivityStrategy extends ASettingsActivityStrategy {
@@ -15,13 +13,13 @@ public class SettingsActivityStrategy extends ASettingsActivityStrategy {
 
 
     private static final String TAG = ".SettingsStrategy";
-    LogoutAndLoginRequiredOnPreferenceClickListener loginRequiredOnPreferenceClickListener;
+    LoginRequiredOnPreferenceClickListener loginRequiredOnPreferenceClickListener;
 
     public SettingsActivityStrategy(SettingsActivity settingsActivity) {
         super(settingsActivity);
 
         loginRequiredOnPreferenceClickListener =
-                new LogoutAndLoginRequiredOnPreferenceClickListener(
+                new LoginRequiredOnPreferenceClickListener(
                         settingsActivity);
 
         pullRequiredOnPreferenceChangeListener = new PullRequiredOnPreferenceChangeListener();
@@ -38,20 +36,6 @@ public class SettingsActivityStrategy extends ASettingsActivityStrategy {
 
     @Override
     public void setupPreferencesScreen(PreferenceScreen preferenceScreen) {
-    }
-
-    @Override
-    public boolean onPreferenceClick(final Preference preference) {
-        if (!settingsActivity.getIntent().getBooleanExtra(SettingsActivity.IS_LOGIN_DONE, false)) {
-
-            String orgUnitValue = settingsActivity.autoCompleteEditTextPreference.getText();
-
-            Intent loginIntent = new Intent(settingsActivity, LoginActivity.class);
-            loginIntent.putExtra(LoginActivity.PULL_REQUIRED, orgUnitValue.isEmpty());
-
-            settingsActivity.startActivity(loginIntent);
-        }
-        return true;
     }
 
     @Override
