+++ app/src/main/java/org/eyeseetea/malariacare/domain/boundary/IAuthenticationManager.java
