+++ app/src/main/java/org/eyeseetea/malariacare/strategies/ASettingsActivityStrategy.java
@@ -8,6 +8,8 @@ import org.eyeseetea.malariacare.SettingsActivity;
 
 public abstract class ASettingsActivityStrategy {
 
+    protected static final String TAG = ".SettingsActivity";
+
     protected SettingsActivity settingsActivity;
 
     public ASettingsActivityStrategy(SettingsActivity settingsActivity) {
@@ -23,6 +25,4 @@ public abstract class ASettingsActivityStrategy {
     public abstract Preference.OnPreferenceClickListener getOnPreferenceClickListener();
 
     public abstract Preference.OnPreferenceChangeListener getOnPreferenceChangeListener();
-
-    public abstract boolean onPreferenceClick(final Preference preferenceScreen);
 }
