+++ app/src/myanmar/java/org/eyeseetea/malariacare/layout/listeners/LogoutAndLoginRequiredOnPreferenceClickListener.java
@@ -1,10 +1,14 @@
 package org.eyeseetea.malariacare.layout.listeners;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
+import android.preference.DialogPreference;
 import android.preference.Preference;
 import android.util.Log;
 
 import org.eyeseetea.malariacare.LoginActivity;
+import org.eyeseetea.malariacare.R;
 import org.eyeseetea.malariacare.SettingsActivity;
 import org.eyeseetea.malariacare.data.authentication.AuthenticationManager;
 import org.eyeseetea.malariacare.domain.usecase.LogoutUseCase;
@@ -20,28 +24,42 @@ public class LogoutAndLoginRequiredOnPreferenceClickListener implements
     /**
      * Reference to the activity so you can use this from the activity or the fragment
      */
-    SettingsActivity activity;
+    SettingsActivity settingsActivity;
 
     public LogoutAndLoginRequiredOnPreferenceClickListener(SettingsActivity activity) {
-        this.activity = activity;
+        this.settingsActivity = activity;
     }
 
     @Override
     public boolean onPreferenceClick(final Preference preference) {
-        return activity.onPreferenceClick(preference);
+        new AlertDialog.Builder(settingsActivity)
+                .setTitle(settingsActivity.getString(R.string.app_logout))
+                .setMessage(settingsActivity.getString(R.string.settings_menu_logout_message))
+                .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface arg0, int arg1) {
+                        logout();
+                    }
+                })
+                .setNegativeButton(android.R.string.no, new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+                        ((DialogPreference) preference).getDialog().dismiss();
+                        dialog.cancel();
+                    }
+                }).create().show();
+        return true;
     }
 
     public void logout() {
         Log.d(TAG, "Logging out...");
-        AuthenticationManager authenticationManager = new AuthenticationManager(activity);
+        AuthenticationManager authenticationManager = new AuthenticationManager(settingsActivity);
         LogoutUseCase logoutUseCase = new LogoutUseCase(authenticationManager);
 
         logoutUseCase.execute(new LogoutUseCase.Callback() {
             @Override
             public void onLogoutSuccess() {
-                Intent loginIntent = new Intent(activity, LoginActivity.class);
-                activity.finish();
-                activity.startActivity(loginIntent);
+                Intent loginIntent = new Intent(settingsActivity, LoginActivity.class);
+                settingsActivity.finish();
+                settingsActivity.startActivity(loginIntent);
             }
 
             @Override
