+++ app/src/myanmar/java/org/eyeseetea/malariacare/strategies/SettingsActivityStrategy.java
@@ -1,9 +1,6 @@
 package org.eyeseetea.malariacare.strategies;
 
 
-import android.app.AlertDialog;
-import android.content.DialogInterface;
-import android.preference.DialogPreference;
 import android.preference.Preference;
 import android.preference.PreferenceCategory;
 import android.preference.PreferenceScreen;
@@ -14,12 +11,12 @@ import org.eyeseetea.malariacare.layout.listeners.LogoutAndLoginRequiredOnPrefer
 
 public class SettingsActivityStrategy extends ASettingsActivityStrategy {
 
-    LogoutAndLoginRequiredOnPreferenceClickListener loginRequiredOnPreferenceClickListener;
+    LogoutAndLoginRequiredOnPreferenceClickListener logoutAndloginRequiredOnPreferenceClickListener;
 
     public SettingsActivityStrategy(SettingsActivity settingsActivity) {
         super(settingsActivity);
 
-        loginRequiredOnPreferenceClickListener =
+        logoutAndloginRequiredOnPreferenceClickListener =
                 new LogoutAndLoginRequiredOnPreferenceClickListener(settingsActivity);
     }
 
@@ -44,30 +41,11 @@ public class SettingsActivityStrategy extends ASettingsActivityStrategy {
 
     @Override
     public Preference.OnPreferenceClickListener getOnPreferenceClickListener() {
-        return loginRequiredOnPreferenceClickListener;
+        return logoutAndloginRequiredOnPreferenceClickListener;
     }
 
     @Override
     public Preference.OnPreferenceChangeListener getOnPreferenceChangeListener() {
         return null;
-    }
-
-    @Override
-    public boolean onPreferenceClick(final Preference preference) {
-        new AlertDialog.Builder(settingsActivity)
-                .setTitle(settingsActivity.getString(R.string.app_logout))
-                .setMessage(settingsActivity.getString(R.string.settings_menu_logout_message))
-                .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface arg0, int arg1) {
-                        loginRequiredOnPreferenceClickListener.logout();
-                    }
-                })
-                .setNegativeButton(android.R.string.no, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int which) {
-                        ((DialogPreference) preference).getDialog().dismiss();
-                        dialog.cancel();
-                    }
-                }).create().show();
-        return true;
     }
 }
